client/requirements.txt
grpcio
grpcio-tools
protobuf
numpy
opencv-python-headless
matplotlib
ffmpeg-python